pre-commit lint over the entire library
