+++ README.md
@@ -146,4 +146,3 @@ Saasha Joshi
 This project is licensed under the Apache License - see the [LICENSE](https://github.com/SaashaJoshi/quantum-image-processing/blob/main/LICENSE) file for details.
 
 ## Acknowledgments
-
