+++ piqture/embeddings/image_embedding.py
@@ -11,8 +11,10 @@
 """Abstract Base Class for Image Embedding"""
 
 from __future__ import annotations
-from abc import ABC, abstractmethod
+
 import math
+from abc import ABC, abstractmethod
+
 import numpy as np
 from qiskit.circuit import ParameterVector
