+++ piqture/embeddings/image_embeddings/frqi.py
@@ -11,12 +11,13 @@
 """Flexible Representation of Quantum Images (FRQI)"""
 
 from __future__ import annotations
+
 import math
+
 import numpy as np
 from qiskit.circuit import QuantumCircuit
-from piqture.embeddings.image_embedding import (
-    ImageEmbedding,
-)
+
+from piqture.embeddings.image_embedding import ImageEmbedding
 from piqture.mixin.image_embedding_mixin import ImageMixin
 
 
@@ -33,7 +34,7 @@ class FRQI(ImageEmbedding, ImageMixin):
 
         # FRQI circuit
         self._circuit = QuantumCircuit(self.feature_dim + 1)
-        self.qr = self._circuit.qubits
+        self.q_reg = self._circuit.qubits
 
     @property
     def circuit(self):
