+++ piqture/gates/unitary_block.py
@@ -11,6 +11,7 @@
 """Unitary Gate class"""
 
 from __future__ import annotations
+
 from abc import ABC, abstractmethod
