+++ piqture/neural_networks/__init__.py
@@ -10,7 +10,7 @@
 
 """Neural Network Structures"""
 
-from .quantum_neural_network import QuantumNeuralNetwork
 from .qcnn import QCNN
+from .quantum_neural_network import QuantumNeuralNetwork
 
 __all__ = ["QuantumNeuralNetwork", "QCNN"]
