+++ piqture/tensor_networks/base_tensor_network.py
@@ -11,7 +11,9 @@
 """Base Tensor Network Circuit"""
 
 from __future__ import annotations
+
 from abc import ABC
+
 from qiskit.circuit import QuantumCircuit
 
 # pylint: disable=too-few-public-methods
