+++ piqture/tensor_networks/ttn.py
@@ -11,13 +11,14 @@
 """Tree Tensor Network (TTN)"""
 
 from __future__ import annotations
+
 from typing import Callable
+
 import numpy as np
-from qiskit.circuit import QuantumCircuit, ParameterVector
+from qiskit.circuit import ParameterVector, QuantumCircuit
+
 from piqture.gates.two_qubit_unitary import TwoQubitUnitary
-from piqture.tensor_networks.base_tensor_network import (
-    BaseTensorNetwork,
-)
+from piqture.tensor_networks.base_tensor_network import BaseTensorNetwork
 
 
 class TTN(BaseTensorNetwork):
