+++ tests/neural_networks/layers/test_base_layer.py
@@ -11,12 +11,12 @@
 """Tests for Base Layer abstract class"""
 
 from __future__ import annotations
+
 import pytest
 from pytest import raises
 from qiskit.circuit import QuantumCircuit
-from piqture.neural_networks.layers import (
-    QuantumConvolutionalLayer,
-)
+
+from piqture.neural_networks.layers import QuantumConvolutionalLayer
 
 
 class TestQuantumConvolutionalLayer:
