+++ tests/neural_networks/layers/test_pooling_layer.py
@@ -11,13 +11,12 @@
 """Tests for Quantum Pooling Layer classes"""
 
 from __future__ import annotations
+
 import pytest
 from pytest import raises
 from qiskit.circuit import QuantumCircuit
-from piqture.neural_networks.layers import (
-    QuantumPoolingLayer2,
-    QuantumPoolingLayer3,
-)
+
+from piqture.neural_networks.layers import QuantumPoolingLayer2, QuantumPoolingLayer3
 
 # pylint: disable=not-context-manager
